+++ src/rules/no-named-default.js
@@ -1,0 +1,19 @@
+module.exports = {
+  meta: {
+    docs: {},
+  },
+
+  create: function (context) {
+    return {
+      'ImportDeclaration': function (node) {
+        node.specifiers.forEach(function (im) {
+          if (im.type === 'ImportSpecifier' && im.imported.name === 'default') {
+            context.report({
+              node: im.local,
+              message: `Use default import syntax to import \'${im.local.name}\'.` })
+          }
+        })
+      },
+    }
+  },
+}